Added random walk

+++ motstat/plotting.py
@@ -119,7 +119,7 @@ class PlotterHist:
         )
 
 
-    def add_lin_segments_hist(self, lin_segments_duration_idxs: List[int]):
+    def add_lin_segments_hist(self, lin_segments_duration_idxs: List[float]):
         trace = go.Histogram(x=lin_segments_duration_idxs, xbins=dict(size=1), histnorm='percent')
         self.fig.add_trace(trace)
